pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gr8academy"
version = "0.1.0"
description = "AI-powered personalized study platform"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Gr8Academy Team" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "bcrypt>=4.2.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "qdrant-client>=1.12.0",
    "openai>=1.58.0",
    "pypdf>=5.1.0",
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
    "Pillow>=11.0.0",
    "python-multipart>=0.0.18",
    "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
